requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
yt-dlp==2024.12.23
google-api-python-client==2.160.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
